Remove superfluous boostrap code
The bootstrap code calls some methods on TSFE that are not needed. The template parsing is done in "getConfigArray()" either way, rootline building in "determineId()".
On my test installation, the original code did not correctly parse the TypoScript templates and as a consequence the correct plugin.tx_rest configuration was not loaded. Removing the superfluous calls fixed the problem.