No access control on OPTIONS method
The OPTIONS method is neither a read nor write
operation. When used for preflight CORS header
discovery no auth can be sent.

This solution changed the isWrite method to
assert that an OPTIONS request is not a write
request. Further the ConfigurationBasedAccessController
is changed to allow requests are neither read
nor write.