Request::getSentData can be called twice
In the case of application/json data being PUT or PATCHED Request was using
(string)$this->getBody() to fetch the payload for json_decoding. Unfortunately
this consumes the Stream and subsequent calls return nothing.

With this commit the sentData is cached in the Request object and returned on
subsequent calls